Current repo
ws.zip contains all the code we currently have. the file "run" has the arduino code that should be flashed to the arduino. all other code should be able to be run in ros; everything in the /dev folder can be moved to the home directory, and run from their respective folders.